requirements.txt
PyYAML>=6.0
uuid6>=2024.1.25
